% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrb_download.R
\name{bmrb_download}
\alias{bmrb_download}
\title{bmrb_download}
\usage{
bmrb_download(id_list, output_dir,
  base_url = "http://www.bmrb.wisc.edu/ftp/pub/bmrb/entry_lists/nmr-star3.1",
  verbose = TRUE)
}
\arguments{
\item{id_list}{(Required) A list of file ids that are corresponding to the requested files.}

\item{output_dir}{(Required) Location were downloaded file will be saved.}

\item{base_url}{Default location is http://www.bmrb.wisc.edu/ftp/pub/bmrb/entry_lists/nmr-star3.1/.}

\item{verbose}{Boolean parameter. If set to be "True", the downloader will output detailed results in the console.}
}
\value{
Save file in the output_dir location
}
\description{
Main function that download the BMRB file from www.bmrb.wisc.edu repo. It will download file in nmr-star3.1 format.
}
\examples{
download_dir = tempdir()
bmrb_download(965, download_dir)
}
