% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_ids.R
\name{collect_ids}
\alias{collect_ids}
\title{collect_ids}
\usage{

  collect_ids(base_url = "http://www.bmrb.wisc.edu/ftp/pub/bmrb/entry_lists/nmr-star3.1/",
  to_list = FALSE)
}
\arguments{
\item{base_url}{(optional) The BMRB entry list page for nmr-star3.1, http://www.bmrb.wisc.edu/ftp/pub/bmrb/entry_lists/nmr-star3.1/.}

\item{to_list}{(optional) whether to output as a list of ids.}
}
\value{
BMRB_files. This could be a list of ids if output, if 'to_list' is set to be True, otherwise, it will return a html table.
}
\description{
Function will parse all the files of BMRB nmr-star 3.1 repo and return all the available files that are available for downloading.
}
\examples{
# collect_ids(to_list=TRUE) # It will take more than 5 sec
}
